+++ e2e/tests/core/02-client/client_test.go
@@ -357,8 +357,8 @@ func (s *ClientTestSuite) TestClient_Update_Misbehaviour() {
 		tmClientState, ok := clientState.(*ibctm.ClientState)
 		s.Require().True(ok)
 
-		trustedHeight, ok = tmClientState.GetLatestHeight().(clienttypes.Height)
-		s.Require().True(ok)
+		trustedHeight = tmClientState.LatestHeight
+		s.Require().True(trustedHeight.GT(clienttypes.ZeroHeight()))
 	})
 
 	t.Run("update clients", func(t *testing.T) {
@@ -373,8 +373,8 @@ func (s *ClientTestSuite) TestClient_Update_Misbehaviour() {
 		tmClientState, ok := clientState.(*ibctm.ClientState)
 		s.Require().True(ok)
 
-		latestHeight, ok = tmClientState.GetLatestHeight().(clienttypes.Height)
-		s.Require().True(ok)
+		latestHeight = tmClientState.LatestHeight
+		s.Require().True(latestHeight.GT(clienttypes.ZeroHeight()))
 	})
 
 	t.Run("create validator set", func(t *testing.T) {