+++ modules/light-clients/06-solomachine/client_state_test.go
@@ -850,7 +850,7 @@ func (suite *SoloMachineTestSuite) TestGetTimestampAtHeight() {
 		{
 			name:        "get timestamp at height exists",
 			clientState: suite.solomachine.ClientState(),
-			height:      suite.solomachine.ClientState().GetLatestHeight(),
+			height:      clienttypes.NewHeight(0, suite.solomachine.ClientState().Sequence),
 			expValue:    suite.solomachine.ClientState().ConsensusState.Timestamp,
 			expPass:     true,
 		},