+++ modules/light-clients/08-wasm/light_client_module.go
@@ -1,0 +1,288 @@
+package wasm
+
+import (
+	errorsmod "cosmossdk.io/errors"
+
+	sdk "github.com/cosmos/cosmos-sdk/types"
+
+	wasmkeeper "github.com/cosmos/ibc-go/modules/light-clients/08-wasm/keeper"
+	"github.com/cosmos/ibc-go/modules/light-clients/08-wasm/types"
+	clienttypes "github.com/cosmos/ibc-go/v8/modules/core/02-client/types"
+	ibcerrors "github.com/cosmos/ibc-go/v8/modules/core/errors"
+	"github.com/cosmos/ibc-go/v8/modules/core/exported"
+)
+
+var _ exported.LightClientModule = (*LightClientModule)(nil)
+
+// LightClientModule implements the core IBC api.LightClientModule interface.
+type LightClientModule struct {
+	keeper        wasmkeeper.Keeper
+	storeProvider exported.ClientStoreProvider
+}
+
+// NewLightClientModule creates and returns a new 08-wasm LightClientModule.
+func NewLightClientModule(keeper wasmkeeper.Keeper) LightClientModule {
+	return LightClientModule{
+		keeper: keeper,
+	}
+}
+
+// RegisterStoreProvider is called by core IBC when a LightClientModule is added to the router.
+// It allows the LightClientModule to set a ClientStoreProvider which supplies isolated prefix client stores
+// to IBC light client instances.
+func (l *LightClientModule) RegisterStoreProvider(storeProvider exported.ClientStoreProvider) {
+	l.storeProvider = storeProvider
+}
+
+// Initialize unmarshals the provided client and consensus states and performs basic validation. It calls into the
+// clientState.Initialize method.
+//
+// CONTRACT: clientID is validated in 02-client router, thus clientID is assumed here to have the format 08-wasm-{n}.
+func (l LightClientModule) Initialize(ctx sdk.Context, clientID string, clientStateBz, consensusStateBz []byte) error {
+	var clientState types.ClientState
+	if err := l.keeper.Codec().Unmarshal(clientStateBz, &clientState); err != nil {
+		return err
+	}
+
+	if err := clientState.Validate(); err != nil {
+		return err
+	}
+
+	var consensusState types.ConsensusState
+	if err := l.keeper.Codec().Unmarshal(consensusStateBz, &consensusState); err != nil {
+		return err
+	}
+
+	if err := consensusState.ValidateBasic(); err != nil {
+		return err
+	}
+
+	clientStore := l.storeProvider.ClientStore(ctx, clientID)
+	cdc := l.keeper.Codec()
+
+	return clientState.Initialize(ctx, cdc, clientStore, &consensusState)
+}
+
+// VerifyClientMessage obtains the client state associated with the client identifier and calls into the clientState.VerifyClientMessage method.
+//
+// CONTRACT: clientID is validated in 02-client router, thus clientID is assumed here to have the format 08-wasm-{n}.
+func (l LightClientModule) VerifyClientMessage(ctx sdk.Context, clientID string, clientMsg exported.ClientMessage) error {
+	clientStore := l.storeProvider.ClientStore(ctx, clientID)
+	cdc := l.keeper.Codec()
+
+	clientState, found := types.GetClientState(clientStore, cdc)
+	if !found {
+		return errorsmod.Wrap(clienttypes.ErrClientNotFound, clientID)
+	}
+
+	return clientState.VerifyClientMessage(ctx, l.keeper.Codec(), clientStore, clientMsg)
+}
+
+// CheckForMisbehaviour obtains the client state associated with the client identifier and calls into the clientState.CheckForMisbehaviour method.
+//
+// CONTRACT: clientID is validated in 02-client router, thus clientID is assumed here to have the format 08-wasm-{n}.
+func (l LightClientModule) CheckForMisbehaviour(ctx sdk.Context, clientID string, clientMsg exported.ClientMessage) bool {
+	clientStore := l.storeProvider.ClientStore(ctx, clientID)
+	cdc := l.keeper.Codec()
+
+	clientState, found := types.GetClientState(clientStore, cdc)
+	if !found {
+		panic(errorsmod.Wrap(clienttypes.ErrClientNotFound, clientID))
+	}
+
+	return clientState.CheckForMisbehaviour(ctx, cdc, clientStore, clientMsg)
+}
+
+// UpdateStateOnMisbehaviour obtains the client state associated with the client identifier and calls into the clientState.UpdateStateOnMisbehaviour method.
+//
+// CONTRACT: clientID is validated in 02-client router, thus clientID is assumed here to have the format 08-wasm-{n}.
+func (l LightClientModule) UpdateStateOnMisbehaviour(ctx sdk.Context, clientID string, clientMsg exported.ClientMessage) {
+	clientStore := l.storeProvider.ClientStore(ctx, clientID)
+	cdc := l.keeper.Codec()
+
+	clientState, found := types.GetClientState(clientStore, cdc)
+	if !found {
+		panic(errorsmod.Wrap(clienttypes.ErrClientNotFound, clientID))
+	}
+
+	clientState.UpdateStateOnMisbehaviour(ctx, cdc, clientStore, clientMsg)
+}
+
+// UpdateState obtains the client state associated with the client identifier and calls into the clientState.UpdateState method.
+//
+// CONTRACT: clientID is validated in 02-client router, thus clientID is assumed here to have the format 08-wasm-{n}.
+func (l LightClientModule) UpdateState(ctx sdk.Context, clientID string, clientMsg exported.ClientMessage) []exported.Height {
+	clientStore := l.storeProvider.ClientStore(ctx, clientID)
+	cdc := l.keeper.Codec()
+
+	clientState, found := types.GetClientState(clientStore, cdc)
+	if !found {
+		panic(errorsmod.Wrap(clienttypes.ErrClientNotFound, clientID))
+	}
+
+	return clientState.UpdateState(ctx, cdc, clientStore, clientMsg)
+}
+
+// VerifyMembership obtains the client state associated with the client identifier and calls into the clientState.VerifyMembership method.
+//
+// CONTRACT: clientID is validated in 02-client router, thus clientID is assumed here to have the format 08-wasm-{n}.
+func (l LightClientModule) VerifyMembership(
+	ctx sdk.Context,
+	clientID string,
+	height exported.Height,
+	delayTimePeriod uint64,
+	delayBlockPeriod uint64,
+	proof []byte,
+	path exported.Path,
+	value []byte,
+) error {
+	clientStore := l.storeProvider.ClientStore(ctx, clientID)
+	cdc := l.keeper.Codec()
+
+	clientState, found := types.GetClientState(clientStore, cdc)
+	if !found {
+		return errorsmod.Wrap(clienttypes.ErrClientNotFound, clientID)
+	}
+
+	return clientState.VerifyMembership(ctx, clientStore, cdc, height, delayTimePeriod, delayBlockPeriod, proof, path, value)
+}
+
+// VerifyNonMembership obtains the client state associated with the client identifier and calls into the clientState.VerifyNonMembership method.
+//
+// CONTRACT: clientID is validated in 02-client router, thus clientID is assumed here to have the format 08-wasm-{n}.
+func (l LightClientModule) VerifyNonMembership(
+	ctx sdk.Context,
+	clientID string,
+	height exported.Height,
+	delayTimePeriod uint64,
+	delayBlockPeriod uint64,
+	proof []byte,
+	path exported.Path,
+) error {
+	clientStore := l.storeProvider.ClientStore(ctx, clientID)
+	cdc := l.keeper.Codec()
+
+	clientState, found := types.GetClientState(clientStore, cdc)
+	if !found {
+		return errorsmod.Wrap(clienttypes.ErrClientNotFound, clientID)
+	}
+
+	return clientState.VerifyNonMembership(ctx, clientStore, cdc, height, delayTimePeriod, delayBlockPeriod, proof, path)
+}
+
+// Status obtains the client state associated with the client identifier and calls into the clientState.Status method.
+//
+// CONTRACT: clientID is validated in 02-client router, thus clientID is assumed here to have the format 08-wasm-{n}.
+func (l LightClientModule) Status(ctx sdk.Context, clientID string) exported.Status {
+	clientStore := l.storeProvider.ClientStore(ctx, clientID)
+	cdc := l.keeper.Codec()
+
+	clientState, found := types.GetClientState(clientStore, cdc)
+	if !found {
+		return exported.Unknown
+	}
+
+	return clientState.Status(ctx, clientStore, cdc)
+}
+
+// LatestHeight returns the latest height for the client state for the given client identifier.
+// If no client is present for the provided client identifier a zero value height is returned.
+//
+// CONTRACT: clientID is validated in 02-client router, thus clientID is assumed here to have the format 08-wasm-{n}.
+func (l LightClientModule) LatestHeight(ctx sdk.Context, clientID string) exported.Height {
+	clientStore := l.storeProvider.ClientStore(ctx, clientID)
+	cdc := l.keeper.Codec()
+
+	clientState, found := types.GetClientState(clientStore, cdc)
+	if !found {
+		return clienttypes.ZeroHeight()
+	}
+
+	return clientState.LatestHeight
+}
+
+// TimestampAtHeight obtains the client state associated with the client identifier and calls into the clientState.GetTimestampAtHeight method.
+//
+// CONTRACT: clientID is validated in 02-client router, thus clientID is assumed here to have the format 08-wasm-{n}.
+func (l LightClientModule) TimestampAtHeight(ctx sdk.Context, clientID string, height exported.Height) (uint64, error) {
+	clientStore := l.storeProvider.ClientStore(ctx, clientID)
+	cdc := l.keeper.Codec()
+
+	clientState, found := types.GetClientState(clientStore, cdc)
+	if !found {
+		return 0, errorsmod.Wrap(clienttypes.ErrClientNotFound, clientID)
+	}
+
+	return clientState.GetTimestampAtHeight(ctx, clientStore, cdc, height)
+}
+
+// RecoverClient asserts that the substitute client is a wasm client. It obtains the client state associated with the
+// subject client and calls into the subjectClientState.CheckSubstituteAndUpdateState method.
+//
+// CONTRACT: clientID is validated in 02-client router, thus clientID is assumed here to have the format 08-wasm-{n}.
+func (l LightClientModule) RecoverClient(ctx sdk.Context, clientID, substituteClientID string) error {
+	substituteClientType, _, err := clienttypes.ParseClientIdentifier(substituteClientID)
+	if err != nil {
+		return err
+	}
+
+	if substituteClientType != types.Wasm {
+		return errorsmod.Wrapf(clienttypes.ErrInvalidClientType, "expected: %s, got: %s", types.Wasm, substituteClientType)
+	}
+
+	cdc := l.keeper.Codec()
+
+	clientStore := l.storeProvider.ClientStore(ctx, clientID)
+	clientState, found := types.GetClientState(clientStore, cdc)
+	if !found {
+		return errorsmod.Wrap(clienttypes.ErrClientNotFound, clientID)
+	}
+
+	substituteClientStore := l.storeProvider.ClientStore(ctx, substituteClientID)
+	substituteClient, found := types.GetClientState(substituteClientStore, cdc)
+	if !found {
+		return errorsmod.Wrap(clienttypes.ErrClientNotFound, substituteClientID)
+	}
+
+	return clientState.CheckSubstituteAndUpdateState(ctx, cdc, clientStore, substituteClientStore, substituteClient)
+}
+
+// VerifyUpgradeAndUpdateState obtains the client state associated with the client identifier and calls into the clientState.VerifyUpgradeAndUpdateState method.
+// The new client and consensus states will be unmarshaled and an error is returned if the new client state is not at a height greater
+// than the existing client.
+//
+// CONTRACT: clientID is validated in 02-client router, thus clientID is assumed here to have the format 08-wasm-{n}.
+func (l LightClientModule) VerifyUpgradeAndUpdateState(
+	ctx sdk.Context,
+	clientID string,
+	newClient []byte,
+	newConsState []byte,
+	upgradeClientProof,
+	upgradeConsensusStateProof []byte,
+) error {
+	cdc := l.keeper.Codec()
+
+	var newClientState types.ClientState
+	if err := cdc.Unmarshal(newClient, &newClientState); err != nil {
+		return errorsmod.Wrap(clienttypes.ErrInvalidClient, err.Error())
+	}
+
+	var newConsensusState types.ConsensusState
+	if err := cdc.Unmarshal(newConsState, &newConsensusState); err != nil {
+		return errorsmod.Wrap(clienttypes.ErrInvalidConsensus, err.Error())
+	}
+
+	clientStore := l.storeProvider.ClientStore(ctx, clientID)
+	clientState, found := types.GetClientState(clientStore, cdc)
+	if !found {
+		return errorsmod.Wrap(clienttypes.ErrClientNotFound, clientID)
+	}
+
+	// last height of current counterparty chain must be client's latest height
+	lastHeight := clientState.LatestHeight
+	if !newClientState.LatestHeight.GT(lastHeight) {
+		return errorsmod.Wrapf(ibcerrors.ErrInvalidHeight, "upgraded client height %s must be at greater than current client height %s", newClientState.LatestHeight, lastHeight)
+	}
+
+	return clientState.VerifyUpgradeAndUpdateState(ctx, cdc, clientStore, &newClientState, &newConsensusState, upgradeClientProof, upgradeConsensusStateProof)
+}